refactor(types): move the setup-advisor wire tests to a sibling module
The inline test module put the file at 308 lines, over the repo's 300-line
limit. Its sibling modules in mainframe-server already split tests this way
(fingerprint.rs + fingerprint/tests.rs), so follow that rather than leave a
new file over the limit.
